python-walk/pyproject.toml
[tool.poetry]
name = "frameio-s3-archiver"
version = "0.1.0"
description = ""
authors = ["Jeff Hodges <jhodges@frame.io>"]

[tool.poetry.dependencies]
python = "3.9.5"
frameioclient = {git = "https://github.com/Frameio/python-frameio-client.git", rev = "develop"}
fastapi = "^0.68.2"
boto3 = "^1.18.56"
python-dotenv = "^0.19.0"
gunicorn = "^20.1.0"
uvicorn = "^0.15.0"
simplejson = "^3.17.5"
pyngrok = "^5.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
